refactor(core): modularize decorators and add style guide
