--- src/core/decorators.ts
@@ -1,235 +1,0 @@
-/* eslint-disable @typescript-eslint/no-unused-vars */
-/* eslint-disable @typescript-eslint/no-explicit-any */
-// src/core/decorators.ts
-import 'reflect-metadata';
-import { COMPONENT_META_KEY, ComponentMeta } from 'echelon/core/types';
-
-// 헬퍼 함수: 클래스의 메타데이터를 안전하게 가져오거나 새로 생성
-function getOrCreateComponentMeta(targetOrConstructor: any): ComponentMeta {
-  const constructor = typeof targetOrConstructor === 'function' ? targetOrConstructor : targetOrConstructor.constructor;
-  
-  if (!Reflect.hasOwnMetadata(COMPONENT_META_KEY, constructor)) {
-    const componentName = constructor.name || 'AnonymousComponent';
-    const newMeta: ComponentMeta = {
-      componentName,
-      eventHandlers: new Map(),
-      propertyBindings: new Map(),
-      methodBindings: new Map(),
-      styleBindings: new Map(),
-      styleLayoutFields: new Set(),
-      propMappings: new Map(),
-      stateFields: new Set(),
-      storeFields: new Map(),
-      routeParamFields: new Map(),
-      queryParamFields: new Map(),
-      // tagName, renderMethodName 등은 데코레이터에서 직접 설정
-    };
-    Reflect.defineMetadata(COMPONENT_META_KEY, newMeta, constructor);
-    return newMeta;
-  }
-  
-  // 기존 메타데이터가 있다면, 누락된 Set/Map 필드들을 초기화 (방어 코드)
-  const meta = Reflect.getOwnMetadata(COMPONENT_META_KEY, constructor) as ComponentMeta;
-  if (!meta.eventHandlers) meta.eventHandlers = new Map();
-  if (!meta.propertyBindings) meta.propertyBindings = new Map();
-  if (!meta.methodBindings) meta.methodBindings = new Map();
-  if (!meta.styleBindings) meta.styleBindings = new Map();
-  if (!meta.styleLayoutFields) meta.styleLayoutFields = new Set();
-  if (!meta.propMappings) meta.propMappings = new Map();
-  if (!meta.stateFields) meta.stateFields = new Set();
-  if (!meta.storeFields) meta.storeFields = new Map();
-  if (!meta.routeParamFields) meta.routeParamFields = new Map();
-  if (!meta.queryParamFields) meta.queryParamFields = new Map();
-
-  return meta;
-}
-
-// --- 클래스 데코레이터 ---
-export function Component(tagName?: string) {
-  return function <T extends { new (...args: any[]): object }>(constructor: T) {
-    const meta = getOrCreateComponentMeta(constructor);
-    meta.tagName = tagName;
-  };
-}
-
-// --- 메서드 데코레이터 ---
-export function Render() {
-  return function (target: any, propertyKey: string | symbol, descriptor: PropertyDescriptor) {
-    const meta = getOrCreateComponentMeta(target);
-    meta.renderMethodName = propertyKey;
-  };
-}
-
-export function Mounted() {
-  return function (target: any, propertyKey: string | symbol, descriptor: PropertyDescriptor) {
-    const meta = getOrCreateComponentMeta(target);
-    meta.mountedMethodName = propertyKey;
-  };
-}
-
-export function BeforeMount() {
-  return function (target: any, propertyKey: string | symbol, _descriptor: PropertyDescriptor) {
-    const meta = getOrCreateComponentMeta(target);
-    meta.beforeMountMethodName = propertyKey;
-  };
-}
-
-export function BeforeUpdate() {
-  return function (target: any, propertyKey: string | symbol, _descriptor: PropertyDescriptor) {
-    const meta = getOrCreateComponentMeta(target);
-    meta.beforeUpdateMethodName = propertyKey;
-  };
-}
-
-export function Updated() {
-  return function (target: any, propertyKey: string | symbol, _descriptor: PropertyDescriptor) {
-    const meta = getOrCreateComponentMeta(target);
-    meta.updatedMethodName = propertyKey;
-  };
-}
-
-export function BeforeUnmount() {
-  return function (target: any, propertyKey: string | symbol, _descriptor: PropertyDescriptor) {
-    const meta = getOrCreateComponentMeta(target);
-    meta.beforeUnmountMethodName = propertyKey;
-  };
-}
-
-export function ErrorCaptured() {
-  return function (target: any, propertyKey: string | symbol, _descriptor: PropertyDescriptor) {
-    const meta = getOrCreateComponentMeta(target);
-    meta.errorCapturedMethodName = propertyKey;
-  };
-}
-
-export function Destroyed() {
-  return function (target: any, propertyKey: string | symbol, descriptor: PropertyDescriptor) {
-    const meta = getOrCreateComponentMeta(target);
-    meta.destroyedMethodName = propertyKey;
-  };
-}
-
-// 클래스 메서드를 DOM 이벤트 핸들러로 지정
-export function Event(domEventName: string) {
-  return function (target: any, classMethodName: string | symbol, descriptor: TypedPropertyDescriptor<(event: any) => void>) {
-    const meta = getOrCreateComponentMeta(target);
-    meta.eventHandlers.set(classMethodName, domEventName);
-  };
-}
-
-// --- 필드 데코레이터 ---
-export function State() {
-  return function (target: any, classFieldName: string | symbol) {
-    const meta = getOrCreateComponentMeta(target);
-    meta.stateFields.add(classFieldName);
-  };
-}
-
-export function Computed() {
-  return function (target: any, propertyKey: string | symbol, descriptor: PropertyDescriptor) {
-    const meta = getOrCreateComponentMeta(target);
-    if (!meta.computedFields) meta.computedFields = new Set();
-    meta.computedFields.add(propertyKey);
-  };
-}
-
-export function Property(domPropertyName: string) {
-  return function (target: any, classFieldName: string | symbol) {
-    const meta = getOrCreateComponentMeta(target);
-    meta.propertyBindings.set(classFieldName, domPropertyName);
-  };
-}
-
-export function Method(domMethodName: string) {
-  return function (target: any, classFieldName: string | symbol) {
-    const meta = getOrCreateComponentMeta(target);
-    meta.methodBindings.set(classFieldName, domMethodName);
-  };
-}
-
-export function Style(cssStyleName?: string) {
-  return function (target: any, classFieldName: string | symbol) {
-    const meta = getOrCreateComponentMeta(target);
-    const styleName =
-      cssStyleName ||
-      (typeof classFieldName === 'string'
-        ? classFieldName.replace(/([A-Z])/g, '-$1').toLowerCase()
-        : String(classFieldName));
-    meta.styleBindings.set(classFieldName, styleName);
-  };
-}
-
-export function StyleLayout() {
-  return function (target: any, classFieldName: string | symbol) {
-    const meta = getOrCreateComponentMeta(target);
-    meta.styleLayoutFields.add(classFieldName);
-  };
-}
-
-export function Store(id?: string) {
-  return function (target: any, classFieldName: string | symbol) {
-    const meta = getOrCreateComponentMeta(target);
-    const storeId = id || String(classFieldName);
-    if (!meta.storeFields) meta.storeFields = new Map();
-    meta.storeFields.set(classFieldName, storeId);
-    meta.stateFields.add(classFieldName);
-  };
-}
-
-export function Use(storeId: string) {
-  return function (target: any, classFieldName: string | symbol) {
-    const meta = getOrCreateComponentMeta(target);
-    if (!meta.storeFields) meta.storeFields = new Map();
-    meta.storeFields.set(classFieldName, storeId);
-    meta.stateFields.add(classFieldName);
-  };
-}
-
-export function Watch(stateField: string | symbol) {
-  return function (target: any, methodName: string | symbol) {
-    const meta = getOrCreateComponentMeta(target);
-    if (!meta.watchHandlers) meta.watchHandlers = new Map();
-    const arr = meta.watchHandlers.get(stateField) || [];
-    arr.push(methodName);
-    meta.watchHandlers.set(stateField, arr);
-  };
-}
-
-// --- 파라미터 데코레이터 (Render 메서드의 파라미터에 사용) ---
-export function Prop(propNameFromParent: string) {
-  return function (target: any, renderMethodName: string | symbol, parameterIndex: number) {
-    const meta = getOrCreateComponentMeta(target); // target은 클래스의 프로토타입
-    if (meta.renderMethodName && meta.renderMethodName !== renderMethodName) {
-        console.warn(`@Prop used on parameter of method "${String(renderMethodName)}" but @Render is on "${String(meta.renderMethodName)}".`);
-    }
-    meta.propMappings.set(parameterIndex, propNameFromParent);
-  };
-}
-
-export function Children() {
-  return function (target: any, renderMethodName: string | symbol, parameterIndex: number) {
-    const meta = getOrCreateComponentMeta(target);
-    if (meta.renderMethodName && meta.renderMethodName !== renderMethodName) {
-        console.warn(`@Children used on parameter of method "${String(renderMethodName)}" but @Render is on "${String(meta.renderMethodName)}".`);
-    }
-    meta.childrenParamIndex = parameterIndex;
-  };
-}
-export function Param(name: string) {
-  return function (target: any, classFieldName: string | symbol) {
-    const meta = getOrCreateComponentMeta(target);
-    if (!meta.routeParamFields) meta.routeParamFields = new Map();
-    meta.routeParamFields.set(classFieldName, name);
-    meta.stateFields.add(classFieldName);
-  };
-}
-
-export function Query(name: string) {
-  return function (target: any, classFieldName: string | symbol) {
-    const meta = getOrCreateComponentMeta(target);
-    if (!meta.queryParamFields) meta.queryParamFields = new Map();
-    meta.queryParamFields.set(classFieldName, name);
-    meta.stateFields.add(classFieldName);
-  };
-}
-
