+++ src/core/decorators/index.ts
@@ -1,0 +1,5 @@
+export * from './class';
+export * from './lifecycle';
+export * from './dom';
+export * from './state';
+export * from './params';
